refactor(extension): remove legacy VM sandbox loader
Delete the vm.runInContext()/Module._compile based extension loader now
that the native CommonJS loader is validated:
- delete src/util/factory.ts and its sandbox tests
- drop the COC_EXTENSION_LOADER comparison switch, native loading is the
  only production path
- remove sandbox-specific coverage from modules-util tests and add
  load-retry and symlink-entry cases to the native loader tests

Full suite passes with one production loader: native Node.js CommonJS
semantics, no Module.prototype._compile patching, no vm-based extension
execution.
